Moved classRoll to ClassAttendanceList
